Moved functions related to index.html into a separate js file.
