+++ app/index.html
@@ -50,6 +50,6 @@
         </ul>
     </div>
 
-    <script src="js/main.js" async></script>
+    <script src="js/index.js" async></script>
 </body>
 </html>
