+++ app/js/main.js
@@ -89,30 +89,9 @@ var width  = 640,
     btnConfirmOK        = document.querySelector("#confirm-container #btn-OK"),
     btnConfirmCancel    = document.querySelector("#confirm-container #btn-cancel"),
 
-    // Launcher window
-    launcherVersion = document.querySelector("#app-version"),
-
     // Sidebar
     btnDirectoryChange  = document.querySelector("#sidebar #btn-dir-change"),
     btnDirectoryDefault = document.querySelector("#sidebar #btn-dir-default");
-
-/**
- * Occurs when "New Project" is pressed
- */
-function openAnimator() {
-    "use strict";
-    gui.Window.open("animator.html", {
-        position: "center",
-        width: 1050,
-        height: 715,
-        min_width: 590,
-        min_height: 500,
-        toolbar: false,
-        focus: true,
-        icon: "icons/icon.png",
-    });
-    win.close();
-}
 
 /**
  * Occurs when "Main Menu" is pressed
@@ -130,22 +109,6 @@ function openIndex() {
         focus: true,
         icon: "icons/icon.png"
     });
-}
-
-/**
- * Check if we can display the latest news feed
- * and if we cannot, say so.
- */
-function canDisplayNews() {
-    "use strict";
-    var NewsFeed = require("./js/feed"),
-    feed = new NewsFeed({
-      url: "http://charlielee.uk/category/boats-animator/feed/",
-      selectors: {
-        container: document.querySelector(".container-news")
-      }
-    });
-    feed.get();
 }
 
 function startup() {
@@ -1025,13 +988,3 @@ function reload() {
     "use strict";
     win.reloadDev();
 }
-
-/**
- * Get version number from package.json
- */
-fs.readFile("package.json", "utf8", function (err, data) {
-    "use strict";
-    if (err) throw err;
-    var datajsoned = JSON.parse(data);
-    launcherVersion.innerHTML = datajsoned.version;
-});
